+++ FindPOLine/OrderLinesFilters.test.js
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 
 import { OrderLinesFilters } from './OrderLinesFilters';
 
-jest.mock('./ExpenseClassFilter/useExpenseClassOptions', () => ({
+jest.mock('@folio/stripes-acq-components/lib/ExpenseClassFilter/useExpenseClassOptions', () => ({
   useExpenseClassOptions: jest.fn().mockReturnValue([]),
 }));
 
@@ -11,6 +11,7 @@ const renderOrderLinesFilters = () => (render(
   <OrderLinesFilters
     activeFilters={{}}
     onChange={jest.fn}
+    applyFilters={jest.fn}
   />,
 ));
 
@@ -25,7 +26,7 @@ describe('OrderLinesFilters component', () => {
     expect(getByText('stripes-acq-components.filter.acqUnit')).toBeDefined();
     expect(getByText('ui-orders.poLine.acquisitionMethod')).toBeDefined();
     expect(getByText('ui-orders.line.accordion.location')).toBeDefined();
-    expect(getByText('ui-orders.filter.fundCode')).toBeDefined();
+    expect(getByText('stripes-acq-components.filter.fundCode')).toBeDefined();
     expect(getByText('ui-orders.poLine.orderFormat')).toBeDefined();
     expect(getByText('ui-orders.poLine.dateCreated')).toBeDefined();
     expect(getByText('ui-orders.line.accordion.vendor')).toBeDefined();
@@ -33,7 +34,7 @@ describe('OrderLinesFilters component', () => {
     expect(getByText('ui-orders.filter.rush')).toBeDefined();
     expect(getByText('ui-orders.eresource.accessProvider')).toBeDefined();
     expect(getByText('ui-orders.eresource.expectedActivation')).toBeDefined();
-    expect(getByText('ui-orders.filter.expenseClass')).toBeDefined();
+    expect(getByText('stripes-acq-components.filter.expenseClass')).toBeDefined();
     expect(getByText('ui-orders.export.exportDate')).toBeDefined();
   });
 });